move vars and consolidate
